Use verify_request_format! method to check valid request format
This commit fix a problem to be able to render error responses for 
format distinct to html (for example json) because e.request.format is
returning a Mime::Type object that doesn’t match with the values in 
self.class.respond_to that is a hash with keys :html, :json, etc

Using the responders’ verify_requested_format! we are getting exactly
the same goal delegating the tasks in the own responders gem.